add a utility library to support patching glusto for pod support
This module provides the ability to patch the glusto run method
with the ability to run commands in pods. This module implements
the custom run method and a helper class that acts as a context
manager or decorator to patch glusto.

This version adds a log_level argument to oc_rsh in order to
plumb the argument all the way through to the inner glusto.run
method. It also cleans up some the docstrings and variable
names.

Change-Id: I0eeea234b1af5e60e41aae63e46e2d816e8929d2
Signed-off-by: John Mulligan <jmulligan@redhat.com>