Speed up execution of the 'edit_iptables_cns' function
One single run of this func, depending on the size of iptables config
file on remote machine, could take minute or more for execution.
It is completely unacceptable just for updating one single file.
So, replace inline change of remote file with simple single operation of
write to it with whole set of needed inner data.

Having cluster of 2 nodes and trying to run tests which have this func in
its 'setUpClass' method, we could wait up to the 5 minutes just to reach
execution of first test. With growth of amount of cluster nodes this
waiting time would only lineary increase.

Change-Id: Ief0b4d3b879ddc9f65557df54beddbb51bc28ba5