Fix concurrency issue in heketi pod failure test cases
In those tests, we scale down amount of heketi pods to 0 and then
scale it up back to amount of 1. And right after it, we perform request
for list of Heketi PODs, not waiting for Heketi POD appearing in list.
So, fix it by adding waiting for 2 sec which is enough time for Heketi
POD to appear in the POD list.

Change-Id: I5361d5c6eb98ac46b5f49a70edd0a88d550ab0b8