Remove skip of CNS-584 tc
Stop skipping CNS-584 tc with reference to the BZ-1644696 bug.
Instead, add possibility to avoid this "blocker" bug,
which is not the point of test.
The reason is that in kube 3.11 we have "synchronous" delete operation
by default. But we did have it as "async" in all the previous kube
versions. So, add possibility to distinguish kube versions adding
additional "--wait=false" option in kube 1.11+ versions.

Change-Id: Ie92abcf4b67f7237092d62b7e8aa180d9877853c