Make 'wait_for_resource_absence' provide more details when fails
In case we get error when resource is not deleted in time,
we do not have valueable info for debugging in its output.
So, add  safe approach for getting more data such as resource info
and list of it's events. Also, increase default timeout from 2 to 5
minutes and reduce wait step from 10 to 5 seconds.

Change-Id: I99b9946fd0ff13f4b307b28125667ceca98d32c3