Add cmd-runner function which is gluster deployment type agnostic
For the moment, most of the test cases and lib functions in
'openshift_ops' python module support only containerized
Gluster clusters. So, to be able to support standalone Gluster clusters
in addition to the containerized ones, add special function
which allows us to run shell commands on both types of Gluster clusters
autodetecting the type.

As an example, update one of the existing lib functions making
all the existing 'arbiter' test cases work in both cases.

Change-Id: Iad69d7dc692c87eb70ed5caf780cb688f274bf02