Add abstraction between deployment types and gluster commands
Now, it is possible to run glustolibs commands
on both deployment types: containerized and standalone.
It is possible using 'auto_get_gluster_endpoint' as target for
lib function in addition to the '@podcmd.GlustoPod()' decorator.

Example:
"""
    from glustolibs.gluster.volume_ops import get_volume_list

    from cnslibs.common import heketi_libs
    from cnslibs.common import podcmd

    class TestExample(heketi_libs.HeketiBaseClass):
        @podcmd.GlustoPod()
        def test_get_vol_list_either_on_node_or_on_pod(self):
            volumes = get_volume_list('auto_get_gluster_endpoint')
"""
Also, delete all the current usages of 'deployment_type' config option,
and delete the option as unneeded anymore.

Change-Id: I281f287c432a5a9efefda588be436ee285188697