Add events info for 'verify_pvc_status_is_bound' common func
For the moment, we write events of failed PVC only to glusto logs.
And only info of one type of events - 'ProvisioningFailed'.
But in case of CI, it is not really suitable. We need to have info in
the exception message.
So, add all the pvc-related events to the message of raised exception when
PVC fails to reach the "Bound" state.

Change-Id: I60e3b9a86acdf251d282059cd2e494c70ae07c70