Completely rewrite 'heketi_with_device_removal_insuff_space' tc
Because it was very fragile and skipped due to bug.
So, after bug was fixed we need to enable it back.
But, also, considering it's stability, it required complete change of
logic. So, it is rewritten completely and moved to more appropriate
module where we have other test cases which perform device operations.

Change-Id: I1b8c18c8ea883f83861e81e97c534459d7588175