Properly skip TCs when 'additional gluster devices' not configured
For the moment, we have 3 different test cases which use
'additional_devices' config option. All of them will fail in case
of specific structure of config.
So, make those test cases be smart enough to understand absence of
such option data and skip test cases with useful info about it.

Change-Id: Id31de7e20188041d2148edf37b3f8c95645c5e52