Install correct heketi-client version on master nodes
Before, we were installing latest heketi-client version on master nodes.
But, in case we install old OCS/CNS we may end up with false
impression that we have new features, when it is not so.
So, read heketi-client package version from Heketi POD and install
exactly the same version on the master nodes.

Change-Id: Ic4e9edd7f6ef614d28c061712473aaf9b9f652e2