Add server package support to the heketi_version.py module
Before, 'heketi_version.py' module didn't consider version of the
'heketi server' package on the Heketi POD. But it is more valueable than
depend on the client only. So, make all the requested version
comparisons be applied to the server's version.

Interaction with the module's logic stays the same, just one behavioral
thing changes: now, if client version is older than server's one, we
raise exception.

Change-Id: I317102556d4a5f51607c3b9da93da0d4a3b74f7f