Use unique prefixes for test cases in 'test_gluster_services_restart.py'
Right now, we use single prefix for all the test cases of
'test_gluster_services_restart' python test module.
It is fragile approach, because if first test fails to delete
heketi or gluster volume, then all other test cases fail too,
but should not do it.
So, use unique prefixes for each of the test cases in this module,
to make failure be unique per test.

Change-Id: I1cc68ffc0553f79d0b0623335a2aaf985feabade