Fix logic for recreation of SSH connections
Before, in case we had some constant networking issue trying to connect
to a host, we were falling into the endless loop of connection recreation.
The error we got after it is the 'recursion limit exceeded'. It is unexpected.
So, change this logic making it avoid recursion and do just one
connection recreation attempt which is useful to fix broken connections.

Change-Id: Id808edbad7e6d69ad58a75bfbae176fddb173d18