Fix test cases which are failing to delete volume after test completion
In test case when we use large disk size, heketi takes time to create
volume and due to timeout it gives error.
Add fix in such test cases to check if volume created after getting
an exception, get details of such volumes or raise exception in case
it fails to create volume.

Change-Id: I1c23a8c6558c23edf8947771e4f41a4bd3ffd66a
Signed-off-by: vamahaja <vamahaja@redhat.com>