Fix 'test_volume_create_delete_when_block_services_are_down' tc
It happens that regex fail to find expected message in oc events
related to the stopped service. The reason is in the not static
part of the error message, it may vary.
So, fix mentioned tc by removing part of the regex, which can vary.
Also, update several other parts of the tc making it use less codelines.

Change-Id: I905f63b2b9aefbe72df610d223da4569ccb14d5e