Fix str/list/tuple distinguishing in library functions to support py3
We try to distinguish strings from other iterable object types in
several our library functions and now it is not compatible with py3.
The problem is that in py2 strings don't have "__iter__" method,
but it exists in case of py3.
So, fix it by using py2/3 compatible check for string types.

Change-Id: I3f833b5e79f1563a3ae6a2bb9ae43d8919d82c30